backend/pyproject.toml
[tool.poetry]
name = "opscopilot-backend"
version = "0.1.0"
description = "OpsCopilot backend - FastAPI + Postgres + pgvector RAG"
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
openai = "^1.40.0"
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.32"
psycopg = { version = "^3.2.1", extras = ["binary"] }
alembic = "^1.13.2"
python-multipart = "^0.0.9"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
tenacity = "^8.5.0"
pgvector = "^0.3.2"
pypdf = "^5.0.0"
email-validator = "^2.2.0"
sentence-transformers = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
httpx = "^0.27.0"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"